Notification datatable functionality moved to djgentelella
