+++ demo/demoapp/datatables/serializer.py
@@ -5,11 +5,6 @@ from rest_framework import serializers
 from demoapp.models import Person, Country
 from djgentelella.fields.drfdatetime import DateRangeTextWidget, DateTimeRangeTextWidget
 
-from django.contrib.auth.models import User
-from djgentelella.models import Notification
-from django.utils import formats
-from rest_framework.views import exception_handler
-
 
 class PersonFilterSet(FilterSet):
     born_date = DateFromToRangeFilter(
@@ -39,42 +34,6 @@ class PersonSerializer(serializers.ModelSerializer):
 
 class PersonDataTableSerializer(serializers.Serializer):
     data = serializers.ListField(child=PersonSerializer(), required=True)
-    draw = serializers.IntegerField(required=True)
-    recordsFiltered = serializers.IntegerField(required=True)
-    recordsTotal = serializers.IntegerField(required=True)
-
-
-class UserSerializer(serializers.ModelSerializer):
-    id = serializers.IntegerField(read_only=True)
-
-    class Meta:
-        model = User
-        fields = (
-            'id',
-            'username',
-        )
-
-
-class NotificationSerializer(serializers.ModelSerializer):
-    user = UserSerializer()
-    creation_date = serializers.DateTimeField(format=formats.get_format('DATETIME_INPUT_FORMATS')[0])
-
-    class Meta:
-        model = Notification
-        fields = (
-            'message_type',
-            'creation_date',
-            'description',
-            'link',
-            'state',
-            'user',
-            'category',
-            'update_date'
-        )
-
-
-class NotificationDataTableSerializer(serializers.Serializer):
-    data = serializers.ListField(child=NotificationSerializer(), required=True)
     draw = serializers.IntegerField(required=True)
     recordsFiltered = serializers.IntegerField(required=True)
     recordsTotal = serializers.IntegerField(required=True)
