+++ demo/demoapp/tests/unittest/__init__.py
@@ -1,4 +1,3 @@
 from .Calendar import *
 from .StoryLine import *
 from .StoryMap import *
-from .NotificationDatatable import *
