+++ demo/demoapp/urls.py
@@ -7,8 +7,8 @@ from demoapp.views import create_notification_view, bt_modal_display
 from .autocomplete import views as autocompleteviews
 from .calendar.views import calendar_view
 from .chartjs import chart_js_view
-from .datatables.api import PersonViewSet, NotificationViewSet
-from .datatables.views import datatableViewExample, notification_datable_view
+from .datatables.api import PersonViewSet
+from .datatables.views import datatableViewExample
 from .date_range import views as date_ranges
 from .filechunckedupload import views as chunckedupload
 from .formset import add_formset, add_model_formset
@@ -28,7 +28,6 @@ menuclss = MenuItemclass()
 
 router = DefaultRouter()
 router.register('persontableview', PersonViewSet, 'api-persontable')
-router.register('notificationtableview', NotificationViewSet, 'api-notificationtable')
 
 urlpatterns = [
                   path('bt_modal_display', bt_modal_display, name='bt_modal_display'),
@@ -98,6 +97,5 @@ urlpatterns = [
                   path('timeline_view', timeline_view, name="timeline_view"),
                   path('datatable_view', datatableViewExample, name="datatable_view"),
                   path('mediarecord_upload', mediaupload_view, name="mediaupload_view"),
-                  path('notification_datatable_view', notification_datable_view, name='notification_datatable_view'),
                   path('tableapi/', include(router.urls)),
               ] + pclss.get_urls() + countryclss.get_urls() + menuclss.get_urls()
