+++ djgentelella/tests/__init__.py
@@ -2,3 +2,4 @@ from .Calendar_Test import *
 from .StoryLine_Test import *
 from .StoryMap_Test import *
 from .TimeLine_Test import *
+from .NotificationDatatable import *
